Generate and store the API definition
Currently, the API definition for the cameras is split across multiple
files. This commit seeks to put them in a single file, much easier for
generation, with the "yq" tool.

It makes this easier by encoding this knowledge within the "task" tool.

== Design Notes
=== Single Version

This commit only supports a single version of the API, as this is all I
can envisage working with in the near future.