Move away from the generated library approach
While trying to figure out how to generate the command line functions
for the API, it became clear that the API code was ... pretty complex.
It can be made a lot simpler, and more features integrated if it was
manually written.

This, of course, means there's quite a lot more work to do. I probably
will not do most of it — This is a project that I'm using to work my
camera. Still, this is easier for me to reason through.