Implement the Autofocus API
This commit implements the minimal work to make the autofocus API work.
This is just a proof of concept, writing out the skeleton for the API.

== Design Notes
=== No Tests

This commit contains no tests, as I'm about to go for lunch, and this
API is just a pet project for me to do while recording some more stuff.
I should write some, and invariably will whenever I need to debug
something.

=== API Deviation

Interestingly, while the API documents HTTP 200 for this response, the
actual implementation returns a HTTP 204.

It will be interesting to see where else the API deviates!

=== API Versioning

There are two, in principle, places the API might deviate:

1. The major API version, and
2. The camera version

The assumption, for now, is that the API will only ever be *corrected*
via the camera version, and camera versions will need to be updated to
use the API correctly.

Given this, it only makes sense to version the API